requirements.txt
python-telegram-bot==20.7
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3
logging==0.4.9.6
sqlite3